fix(react-pid): fix auto load (#2032)
Unfortunately, I broke master in #2007 . Seems like the issue was here; https://github.com/cognitedata/applications/pull/2007#discussion_r811720316 

Not sure why it never occurred originally while I was developing? 

This is fixed in this branch by letting CognitePid owning the state for symbols and symbolInstances and simply letting the component keep a copy of that state.