feat(react-pid): add panning/zooming with mouse (#2020)
Link to JIRA:
https://cognitedata.atlassian.net/jira/software/projects/PIP/boards/534?selectedIssue=PIP-122&text=zoom

As a labeller, I want to be able to zoom and pan around the workspace with my mouse.

# What has changed?

Adds zooming and panning with mouse/touchpad. 

Removes zooming and panning with keyboard shortcuts - mainly because it's extra effort + I don't think it adds much value once zooming and panning with touchpad/mouse is there.