fix(auth-utils/getFlowKey): always use lowercased key (#2050)
Link to JIRA:

https://cognitedata.atlassian.net/browse/INFIELD-2834 – the issue blocks infield from adopting new login page.

# What has changed?

auth-utils/getFlowKey always gives lowercased value (that impacts both save/read data).

# Any other notes?

That's rather a quick fix, but the issue is that we use localstorage to implicitly communicate between cdf-hub-tenant-selector and react-container's code. The issue happens when user types "Akerbp" (uppercased), cdf-hub-tenant-selector stores a key like `__v3__storage/flow_Akerbp`, but the react-container tries to read it as `__v3__storage/flow_akerbp` – can't find this key, receives that flow is undefined and does nothing (silently redirects back to login page).

I think a better fix would be to remove the part that saves selected flow to localstorage from tenant-selector and save it in `onSubmit` handler instead (on react-container side). But please let me know if there are some reasons to do it the way it's done, cc @vegardok