feat(pid-tools): [Auto analysis] Support line jumps (#2018)
Link to JIRA:
[PIP-64](https://cognitedata.atlassian.net/browse/PIP-64)

# What has changed?

Line jump support. Added a new constant `AUTO_ANALYSIS_LINE_JUMP_THRESHOLD` and it is relatively arbitrary set to 30 (30x higher than regular PID distance threshold). This should however not include that many false positive since it now is a restriction that the edge point on a line only can be connected to one instance. It is also a restriction now that the two lines and the angle of the connection points between the lines cannot be greater than 5 degrees (this is also a constant that can be tweaked).

Did some refactoring of the `pathSegment.angle` function (now returns angle between 0 and 360) and added a `angleDifference` function that makes it easier to work with angles.
 
# Any other notes?

Unnecessary calls to `getDiagramInstanceId` actually had a big performance hit. On RBD_G0040_MF_026-Model.svg the `performance: Don't use getDiagramInstanceId` commit shaved off around 3-4 seconds.