Update eslint
Update eslint to latest available. This include various related
dependencies.

Simplify run-scripts to do linting. The "lint" and "lint:fix"
run-scripts will also include prettier.

I've had to add a few more packages to get things to work. Usually
lots of linting stuff will come from react-scripts, but we have
overridden parts of it and not yet updated it to a compatible
version. We should clean up this later when all dependencies
are up-to-date.

I've disabled some rules to not include too much changes into this
commit, so we might want to review rules later.