fix: package.json repository and homepage (#2117)
Set repository and homepage to point to the correct places.
This fixes some incorrect values, and also adds it to other
packages.

This improves e.g. dependency update processes in other code bases,
as command such as "yarn outdated" will contains a link to the package
code.

I have bumped the patch version on some of the packages, excluding
those that didn't seem to be in use extenally.

---

This PR will likely conflict as soon as any of these packages gets another version bump, so I will likely need to take another round on version numbers for this is merged.