feat(pid): Support different prefixes in line numbers (#2086)
Link to JIRA:
[_insert link here_](https://cognitedata.atlassian.net/jira/software/projects/PIP/boards/534?selectedIssue=PIP-124)

# What has changed
Changed (back) line numbers to be used as a string, including prefix. Updated matching to also detect other prefixes than L, so that it can be used for line number inferring.

To be able to avoid false positives the regex is has beed limited. It does not detect the longer line labels as line numbers, but should support all variations for line breaks.