Fix bug where user was able to exit model config modal without selecting a custom model (#637)
Previously, the user was able to select "Custom Model" in the annotation model dropdown without actually selecting the model itself (from the list of AutoML models returned by the API). This work introduces the necessary changes to make it simpler for the user to avoid this case.

Specifically, this is done by greying out the "Close" and "Save & Close" buttons until the user has selected a custom model (assuming that the user has clicked on the "Add custom model" button). A tooltip is shown when hovered over the "Save & Close" button to guide the user.