feat(pid): upload linewalk files to cdf (#2083)
Link to JIRA:
_insert link here_

# What has changed?
Added python script for uploading parsed files for lineWalk to CDF. This will now be done automatically when running the parser. It now also uploads two new files:

**PARSED_DOCUMENT_VERSIONS.json**
includes a list of uploaded versions. This could be used to implement a version selector in lineWalk for dev purposes. Staging/prod could be locked to latest or a specific stable version.

**PARSED_LINES_V{$version}.json**
All of the lines that are included in the version/revision