refactor(pid-cli): small refactor of linewalk output  generation (#2178)
Small refactor of linewalk output generation. 

Main difference is that there's a separation of concern between generating the data and persisting it. 

Tested with sample data and produces the exact same output (verified with sha checksums). 

Made this change as a precursor for adding "state" data for the Linewalk app.