fix(react-comments): get user id from token if provided (#2177)
# What has changed?

Some lines of code are based on user id taken from authState. For example:
`const isLoading = !authState?.id || !target;`
So if we pass our own token, the component doesn't work as expected.
The idea is to get user id from authState if it is defined. If not, we can extract it from the token itself (see `getUserIdFromToken` function)