feat(pid-cli): add line reviews state (#2188)
This PR does a bit more refactoring. 

Main change is that it produces a line reviews file that serves as really ghetto backend for the Linewalk app. The Linewalk side of this has been built and tested already, but making this PR separately. 

Linewalk file looks something like this;
```
        ({ data: { externalId, line } }) => ({
          id: line,
          name: line,
          system: 'unknown',
          entryFileExternalId: externalId,
          assignees: [{ name: 'Garima' }],
          status: 'OPEN',
          discrepancies: [],
        })
```

Some other notable changes: 
- Stop bumping the version automatically. Instead the version number is passed via a command line argument. I think this makes more sense. 
- The upload script will overwrite files by default.