feat(pid): Insert instances on a clean pid when loading from graph (#2179)
Link to JIRA:
_insert link here_

# What has changed?
When loading a graph from json, the svg is rendered and pidDocument created from scratch to add the graph on a "clean slate". Effectively overwriting any changes in current doc.

The json validation is simplified as it did not make sense to check if all paths exists in the svg before adding the line breaks from the graph. Now expects the line breaks and symbols to match, at it should not be possible to split a path when it is a part of a symbol (see Henriks fix on this in line break PR)

# Other notes
Should probably add a warning "your current changes will be overwritten" at some point

## Testing notes
### Legend
Currently uploading a legend will not overwrite your current legend, and no de-duplication enforced. Meaning you will get duplicates if the legend you upload has any of the symbols already in your legend. 

To test the legend matching "from scratch": 
- download legend when you have a fully labeled document 
- clear local storage legend
- reload document
- import the previously exported legend
- wait a bit
- all symbols should be labeled

### Graph document
- download graph when you have a fully labeled document
- optionally clear local storage legend (will make the symbol instance loading more apparant)
- reload document
- optionally split some lines, label some symbols etc
- import previously exported graph
- should now have the same state as when exported, discarding anything you might have done after the reload step