[VIS-715] Bulk delete annotations (#638)
This work implements _bulk edit of annotations_. The video below shows a demonstration of this feature:

https://user-images.githubusercontent.com/9768053/158635936-e84e9963-1502-4e7d-b22e-223f098a41ba.mov

To achieve this, some changes to the bulk edit modal were made. The most significant one is that we now provide a `rendererMap` to the bulk edit table which then uses that to figure out how to render the columns. Another change that was needed to implement this feature was the introduction of an optional `filter` parameter to the annotation selector. This makes it possible to filter the annotations by, say, the annotation status and/or the label.