[VIS-730] Model settings options in model selector (#663)
This work adds a new option "Edit model settings" option in the model
selector in the process page. The other settings button was a bit vague
and could be difficult to find. Hence, why we add this new select option
to make it easier for the user to find the model settings page.

Co-authored-by: Marvin Reza <marvinreza@Marvins-MacBook-Pro.local>
Co-authored-by: Milad Mobarhan <43236572+miladcognite@users.noreply.github.com>