fix(linewalk): many minor ux tweaks (#2387)
Tweak the appearance of the footnotes in the export
Fix some bugs with regards to “layering” of buttons and modals. 
Improve zooming when clicking through from discrepancy list
Sync the appearance of the discrepancy modal and iso modal with the designs. 
Add close button to the iso modal. 
Improve the drag-ability of the discrepancy modal and the iso modal (larger area is draggable)