feat(pid-cli): add update-metadata command (#2416)
This PR refactors pid-cli to be TS only to keep all of the business logic in one place and adds an `update-metadata` command that tags all the PDFs and parsed documents with metadata. 

# Testing instructions
Run for instance ` yarn cli all --output-version 0.0.18`. Check the output of the CLI as well as the corresponding files in CDF.