feat(pid-tools): Add line breaks to graph matching (#2347)
# What has changed?

This PR adds line breaks to graph matching. Line breaks between PIDs and ISOs are matched if they are unique (only one match) and have the same two lines connected to it.

# Any other notes?

This PR is based on #2314 (for no other reason that is was quicker to label the symbols with that branch). Hence this PR should be reviewed after that

# Step to test

1. Label line breaks in PIDs and ISOs
2. Run multi-graph matching
3. You should see line breaks in Line review now