feat(linewalk): use events as backend (#2440)
Use events as the backend for Linewalk. 

https://cognitedata.atlassian.net/jira/software/projects/PIP/boards/534?selectedIssue=PIP-214

Fix bug when removing all documents in a workspace - this needs more design and has been forwarded to Jonas but for now this should do. 
https://cognitedata.atlassian.net/jira/software/projects/PIP/boards/534?selectedIssue=PIP-217

# Testing
For the pid tools, run 
```yarn cli all```

Then start Linewalk locally `yarn bazel:start` and confirm that there are no regressions :)