refactor(linewalk): handle units (#2479)
PIP-212

This add support for handling units in Linewalk. 

Assumption now is that each document belongs to one unit. 

This might prove to be false - a document might belong to multiple units - or it might turn out that they duplicate the drawings on a per unit basis. I've emailed Reshma to ask how they do it. 

Suggesting we get this merged for now and iterate on it once we know more. 

If we need to support multiple units, we need to modify: 
1. The data structures here to handle multiple values
2. Change how we infer the units (probably based on the labels in the documents?) -> Then we run into the issue of too many documents in the workspace (since ISOs for instance contain references to other units)