[VIS-781] Loading animation on file delete (#703)
Disables the "Bulk action" button and also adds a loading animation when
files are being deleted. This should make it simpler and more explicit
for the user that an operation is ongoing. I decided to also disable the
bulk action button so that the user won't be able to, for example,
perform bulk operations on files that are being deleted.

Specifically, this is achieved by adding a new optional parameter
setIsDeletingState to the DeleteFilesById thunk, which allows the thunk
to set the deletion states before and after deleting is finished.