feat(pid): Parse line numbers with unit (#2608)
Link to JIRA:
[PIP-226](https://cognitedata.atlassian.net/browse/PIP-226)

# What has changed?

This PR adds the correct unit to the different line number labels in the PIDs. I.e. before if PID G0001_MF_01 had line number G0002_L001 and L002, they would be connected to ISOs G0001_L001 and G0001_L002 respectively. Now they should be connected to ISOs G0002_L001 and G0001_L002 instead.

Also the PDF metadata for ISOs will only be linked to the line number in the name. This filtering is done in `useCdfDiagrams`.

# Any other notes?

Some refactoring is made to extract the UNIT_REGEX as a constant.

I removed the feature of clicking on line numbers in the Diagram Parser to "infer line numbers". That flow is not used anymore and we can reintroduce that feature at a later point if we need it.