[VIS-911] Fix error when creating annotation after file delete (#782)
useCallback was using the cached file id since it only reacted to
scrollIntoView. Consequently, when we would attempt to create
annotations after deleting a file, the application would use the old
file id (i.e, the deleted file id) to create the annotation. This would
result in an error with the new API since we now do resource access
checks.

The fix for this was to add `file` to the list of deps to useCallback.
As such, the hook would react correctly on file changes.