[VIS-916] [VIS-927] Fix annotation color in image preview and side panel (#789)
This fixes the bug in the image preview where the annotation color
wasn't set properly. This is done by properly setting the review
annotation color based on the annotationColorMap.

Moreover the colors in the side panels are also fixed. All keypoints within the same keypoint collection are now assigned the same color.