[VIS-894] Fix wrong bulk action selection count (#802)
> **What is the bug?**

This PR fixes a bug where the wrong selection count would be displayed. For example, see the attached image below. Although there only are 4 items in the process page a selection count of 19 is displayed when we click the "select all" check box at the top left.

![image](https://user-images.githubusercontent.com/9768053/175135073-a5fdb6fa-70d1-4b1e-9e40-2b7d0bbd8eee.png)


> **Why does this occur?**

The reason this happens is because the "select all" button in the process page selects all files given by [`fileState.files.allIds`](https://github.com/cognitedata/cdf-vision-subapp/blob/19c7e5aa954bfa025388afcb93efabe8598e0a89/src/store/commonReducers.ts#L42). This is specifically an issue for us in cases where `FetchFilesById` is slow (e.g., due to the Files API being slow) because the fileState will be updated concurrently even if we have switched from the explorer page to the process page. Consequently, `fileState.files.allIds` will be populated with additional file ids that are not shown in the process page.


> **How is this fixed?**

To fix this issue we "simply" modify the selectors `setSelectedAllFiles` and `selectAllFilesSelected` so that we can optionally override them with which file ids we want to select from. This change enables us to only select among the process file ids. Accordingly, we will not be affected by the concurrent changes being made to the fileState. Moreover, this should be what we actually want since it makes sense that we can only select among the files that are visible to the user (namely, the file ids given by `processSlice.fileIds`).