UI Fixes: Don't show visualizer error text in new data model #25a [DX-680] (#492)
* Update GraphQL utils service behavior around empty schema

Formerly, removing all types from a schema would result in the graphql-extra library generating an SDL string of "\n". This is problematic because we try to parse this in our UI components and it throws an error. We now intercept the "\n" string and return "" instead.

* Don't show error in visualizer when schema is empty

* Ensure we update draft when schema becomes empty

fixup

* Show discard button even when schema is empty

Formerly the discard button would not change if you clicked to edit a data model and deleted all the types.

Co-authored-by: cognite-bulldozer[bot] <51074376+cognite-bulldozer[bot]@users.noreply.github.com>