refactor(pid-tools): create msal client with arbitrary credentials + interace with cli (#2960)
Link to JIRA:
[This work is a necessary precursor to this Jira ticket](https://cognitedata.atlassian.net/browse/CXT-684)

# What has changed?

## Extended the `getClient` method and made it more flexible.
Previously, it was possible to use this method to generate a JS client with the Implicit Grant flow, with credentials hardcoded to the `p66-dev` project on `az-eastus-1`. 
Now, there is a more generic `getMsalClient` function which allows to create a client with both, the Implicit Grant flow and the Client Credentials flow (using a client-secret). In addition, it is more generic and takes all necessary argument to log in to any project on any cluster that features Microsoft OIDC flows.

## Updated interfaces of `pid-tools` cli
The `pid-tools` CLIs have been accordingly changed to take those necessary CLI arguments, instead of just assuming the `p66-dev` project. I also took the liberty to clean the CLI code slightly up, mostly removing code duplication in `pid-tools/cli/main.ts`.