feat(discover): a more long term remote fix (#2980)
# What has changed?

Move the layer check into the correct place, and prepare to enable the remote layers again if needed.

The only problem with the new system is the order when it encounters some missing ones. If we want to keep using hybrid layers (eg: remote files and geospatial) then we need to resolve this. 

# Any other notes?

Some minor type cleanup