chore(pid-tools): Clean up SVG Element DOM transformations (#2977)
# What has changed?

The way we have normalized the SVG Element has had a lot of room for improvement. This PR:

- Overall cleans up the API to parse SVG Commands (see that this PR has a significantly negative line number diff)
- Improves the performance:
  - The SVGs `svg.getScreenCTM()?.inverse()` was called on every point before
  - the `svgElement.getScreenCTM()` was called for every point in the path, now it's only called one per SVG path

I hope this would improve the loading time even more -- don't have a very thorough test for this, but it decreased the loading time for 1.4s to 850ms on my test file. I would, however, say that this should mostly be treated as a refactor PR.

## Steps to test

Unfortunately we don't have any automated test to check the functionalities that requires DOM APIs. To do a manual test:

1. Label a symbol. The red bounding box should be places correctly around the symbol.
2. Use the 'Connect Labels' tool to connect a label to a symbol instance. The green bounding box should be correctly placed around the text.