feat(pid-tools): Better symbol symbol to line auto connection (#2989)
Link to JIRA:
[CXT-712](https://cognitedata.atlassian.net/browse/CXT-712)

# What has changed?

Currently the Diagram Parsing Tool is not able to automatically connect many of the instruments to lines due to line gaps. This PR improves this by increasing the threshold it tries to connect between lines and symbols. An additional constraint is added to not introduce too many false positives – that is that the intersection between the mid point of the symbol and line is within negatively padded bounding box of the symbol. The attached picture if probably a lot easier to understand for this:


![Skjermbilde 2022-07-15 kl  17 14 49](https://user-images.githubusercontent.com/7979907/179254649-0b9beb18-8376-4999-ad7b-a898de1fce99.png)
Hence, with this PR the line between the instrument and valve will be connected, while the other line would not be connected.

# Step to test

1. Open a document and label at least one instrument and a symbol it should be connected through with a line with gap.
2. Run 'Auto analysis'
3. Now the instrument will hopefully be connected to other symbol you labelled.

I have noticed that there still are some cases it won't connect automatically. But when I increased the `threshold` many false positives connection were introduced. This can probably be improved in  the future, but I still believe that this is an incremental improve in the correct direction.

# Any other notes?

There are still some problems with the `findLinesAndConnection` functionality

- The constants are not normalized based on the document size
- The algorithm can be agonizingly slow for larger documents — here I am pretty sure that there are some low hanging fruits.