chore(pid-tools): Refactor applying of styles with significant better performance (#3018)
# What has changed?

This PR improves the performance of applying styles quite significantly. On my test file it goes from 1.7s to 340ms when going to the 'Connect instances' from the 'Connect labels' tool and 1.4 to 240ms changing the same tools the other direction — not too bad I would say! This is after one symbol is labeled and 'Auto analysis' is run.

I have iteratively used the 'Performance' tab in Chrome dev tools and tried to work on the part of the code with the biggest impact. A summary is as follow:

- Before we looped through all the `lines` and `symbolInstances` for each path in `applyStyleToNode` in `domUtils.ts`. For some `activeTool` this happed several times. Now a `diagramInstanceId` to `DiagramInstanceWithPath` map is created that is updated on `setSymbolsInstances` and `setLines` in `cognite-pid.ts`.
- Before the different overlays were inserted with a `svg.insertBefore(node, svg.children[0])`. After I changed it to an `appendChild` the performance was a lot better. I assume that the `insertBefore` need to shift all the elements, just as for a list data structure.
- We did do some unnecessary retrieval of node attributes which is relatively expensive. When I changed `originalStrokeWidth` assignment in `drawConnectionVisualizations` in `cognite-pid` it shaved off another 90ms.

I also done some general refactoring, such as changing from `if` / `else if` to `switch` with the `assertNever` utility method (which I stole from [reveal](https://github.com/cognitedata/reveal/blob/master/viewer/packages/utilities/src/assertNever.ts) 😄)

# Steps to test

Everything should be as before, but we don't have any automated tests for these functionalities so please check things manually:

- Label a symbol an check that paths turnes orange and red as before
- Check the 'Add line' tool that it is possible to click on paths and they turn to blue and back to black
- Check that the 'Connect instances' and 'Connect labels' works as before.