feat(pid-tools): First interfaces to upsert DMS nodes (#3012)
Link to JIRA:
https://cognitedata.atlassian.net/browse/CXT-684
Preparation to https://cognitedata.atlassian.net/browse/CXT-701

# What has changed?
## Interfaces to interact with DMS
- First interfaces to upsert engineering-diagram nodes to a DMS space
- For now works for the `Viewbox` type. Upserting and retrieving `SymbolTemplate` type will be the next step.
- Assumes the updated DMS model hierarchy from this PR: https://github.com/cognitedata/matchmakers-property-graphs/pull/11
- Circumvents the current limitations on DMS that upserting of nodes behaves differently from retrieval. Basically the retrieval data structure is assumed, and the data types are sanitized before and after the upserting call to DMS.

## CLI tool to do the upserting
- Works through a CLI interface for now, as this allows rapid prototyping. 
- Loads a file containing a `GraphDocument` (the output format of the parser) and reads the data from there.
- Will be integrated with the main code at a later stage

## How does it work with the DMS type system
DMS has a relatively intricate type system for Nodes, which allows to derive from base node types.
Every node has a "model name" associated with it, and a clear description what properties such a node has (with associated data types).

Mapping the type system itself can be done relatively easily on Typescript: Just create interface for each Node Type, which possibly derive from each other. There is one caveat though: When we upsert a node of a certain type (indicated by the model name string on DMS site), it is not clear on Typescript size to which interface this string maps to. We need such a mapping though in order to cleanly define the argument and return types of the upserting and list methods.

The trick I used here to create such a mapping is to have an additional generic interface `NodeAdapter<N>`, which is defined by a few static properties. Most notably it defines a `modelName` static string property, and also happens to be a generic of a specifc node type `<N>`.  We then create one concrete implementation NodeAdapter class that fulfils this interface for each Node Type. The `ViewboxNode` interface for example corresponds to the `ViewboxNodeAdapter`. The Adpater class fully identifies the `modelName` as well as the interface it maps to. To upsert a `ViewboxNode` we can then call:

`const upsertedNodes = await upsertNodes(client, {
    NodeAdapterType: ViewboxNodeAdapter,
    items: [viewboxNode],
    spaceExternalId,
  });`

The argument `items` will be automatically of type `ViewboxNode[]`, as is the return value `upsertedNodes`. 

## How it works around DMS limitations
DMS currently has the limitation / bug that the upsert data format differs from the retrieval data format for certain properties. To workaround this, we basically assume that we always have the retrieval data format on the Typescript side. Before we upsert, we call a `sanitizeBeforeUpsert` function that transforms a node in the retrieval data format into the upsert data format. When the nodes are returned from the DMS API (still in upsert data format), we cann a `sanitizeAfterUpsert` function to transform it into the retrieval format. This hides the complexity of different data formats from the caller.

Since we already have the `NodeAdapter<N>` interface, those functions (specific for each node type) can be conveniently placed on the Adapter classes that implement this interface.

## How to test
Run 
```
yarn run cli upsert-symbols \
--base-url="https://greenfield.cognitedata.com" \
--project="atlas-greenfield" \
--client-id="ec3684fd-5517-45c1-8a13-1a925b000ab8" \
--tenant-id="0b320d24-2c6f-45c6-be28-8d63e0804f51" \
--client-secret=INSERT_YOUR_ATLAS_GREENFIELD_CLIENT_SECRET_HERE \
--path=PATH_TO_A_VALID_JSON_FILE_CONTAINING_GRAPH_DOCUMENT \
--space-external-id="matchmakers-test-2-engineering-diagrams" \
--file-id=1 \
--file-page=1
```
This should upsert a `ViewboxNode` and afterwards immediately retrieve it, printing the result of the upserting and retrieval into console.