fix(discover): ppfg filters select issue [PP-3092] (#3062)
Link to JIRA:
https://cognitedata.atlassian.net/browse/PP-3092

# What has changed?

Fixed PPFG filters select issue.

It was due to a rerendering issue. As soon as the option selection is toggled, the useEffect triggers and initializes the selection state. The user can't see even the selection. Using the deep hook prevents unnecessary rerendering and the selection works fine after that.