feat(pid-tools): Added unassign asset button and functionality (#3095)
Link to JIRA:
https://cognitedata.atlassian.net/browse/CXT-732

# What has changed?
Previously there was no way to unassign/remove an asset assigned to a symbol instance, you could only re-assign the symbol instance to a new asset. This can be quite annoying and confusing as it is possible to assign the wrong asset to a symbol instance, and there does not always exist a correct alternative. 

- This PR adds a cross icon at the right side of the asset name, if pressed, both the assetName and assetId gets deleted from the symbol instance object, meaning that the asset gets unassigned from the symbol instance.
- The cross also only appears if the symbol instance has an asset assigned to it, so be sure to check this functionality on the green symbol instances after running Auto Analysis in the Connect Labels mode.