feat(pid-tools): Added automatic state loading from CDF in Diagram Parsing Tool (#3086)
Link to JIRA:
https://cognitedata.atlassian.net/browse/CXT-722

# What has changed?
The Diagram Parser State file (aka. GraphDocument) is saved to CDF using the “Save” button or in the “batch processing” mode. This state is, however, not loaded if you load the same document again in the Diagram Parsing Tool. In this PR we would like to automatically load this state file from CDF when a document is loaded (so the user can continue labeling).

# Any other notes?
I did not find any file examples where there was not already created a corresponding json state file, this made it difficult to test for situations in which there is no saved json state. This should absolutely be tested!