feat(pid-tools): Added a deselect rectangle feature (#3094)
Link to JIRA:
https://cognitedata.atlassian.net/browse/CXT-731

# What has changed?
This PR adds a blue deselection rectangle feature, which is enabled just the same way as the red selection rectangle, except that you have to press the option/alt key as well. You can press the option/alt key whenever you want in the process (see the colour changes dynamically ), the only thing that matters is if the key is pressed or not when releasing the mouse button. 

# Any other notes?
Since the pressing of the alt key created a conflict with the already existing alt-key feature, some changes had to applied. The previous deselect alt-key feature is now only enabled if the mouse is not pressed down `if (mouseEvent.altKey && !this.mouseDown)`.