(pid-tools): Show last parsed in list modal in Diagram Parser (#3159)
# What has changed?

This PR adds a "Last parsed" column to the diagram list in the Diagram Parser. The CDF file meta data is also updated slightly, now the SVG file metadata has the following fields:

```typescript
{
  [DIAGRAM_PARSER_JSON_EXTERNAL_ID]: jsonExternalId,
  [DIAGRAM_PARSER_LAST_PARSED_KEY]: `${Date.now()}`,
}
```

<img width="1326" alt="Skjermbilde 2022-08-23 kl  15 58 05" src="https://user-images.githubusercontent.com/7979907/186177539-e170963e-b160-4c82-9ebf-dfe7fd259a88.png">

A (not very pretty) checkbox is also added to **not** load the state from CDF. This is an important feature for development since the state from CDF updates the symbol library to the symbol library saved for the given file. Long-term this toggle should be removed.
<img width="471" alt="Skjermbilde 2022-08-23 kl  16 04 27" src="https://user-images.githubusercontent.com/7979907/186179010-75dfb369-2d91-4930-8914-dace68a91eb1.png">

# Steps to test

- Label a document and click "Save"
- Open the Diagram List modal (a reload might be required)
  - You should now see the "Last parsed" populated for the file you saved with the correct timestamp
- Now click on the "Load state from CDF" toggle (so it is in the off state)
  - You should now **not** load the state you saved previously