[VIS-937] Fix: Don't change annotation confidence on update (#887)
Ideally there should be some sort of versioning of the annotations in case they are created by a model but later modified by the user so that we don't loss information that may be relevant for future ML model training. However, we don't have a concept/solution for this yet, but still we can limit the modification of the original annotation, by not implicitly changing the confidence.

Also [require annotations:write in process page](https://github.com/cognitedata/cdf-vision-subapp/commit/0c4950f78256453bb62aeb422b77e4fec2ea8ba1). Since models will attempt to create annotations on completion,
we can  just as well require this capability from the start.