fix(pid-tool): Speed up automated asset assignment in Diagram Parsing Tool (#3098)
Link to JIRA:
https://cognitedata.atlassian.net/browse/CXT-733

# What has changed?
The previous implementation of the automation used a for/of loop to solve the 429 Too Many Requests Error, meaning that none of the iterations could run in parallel. This is a bit overkill, since we only need a small time offset between the different request, the iterations should still run in parallel. In this PR we try solving this issue with a forEach loop instead that allows for parallelism, and use a `instanceStatusArray` that keeps track of if the nested Promises have been resolved. When all the values in the `instanceStatusArray` are true, the function continues.

I did some tests to see how much this sped up the process of assigning assets to 'ASDAA-04-00008-0001.svg' by averaging over 3 runs:

Old for/of average time: **17906ms**
New forEach average time: **2919ms**

As you can see, this PR should improve the speed by **6.13 times**. 

# Any other notes?
I ran through the whole batch without any 429s, should they still occur, simply increase the timeout which is now set to 25ms.