(pid-tools): Fix auto analysis when a path only has one CurveSegment (#3237)
BUGFIX:

# What has changed?

This PR fixes a bug when a document has a path with only one curve segment and the "Auto analysis" feature is clicked (it hangs forever). This would for instance happen in the file "ASDAA-04-00010-0001.svg". This happened due to the old throw expression in line 127.

Now this check is handled outside of the `getCloseWithLineJumps` so getClosePidInstances will take another branch.