feat(pid-tools): groundwork for upserting symbol instances (#3042)
Link to JIRA:
https://cognitedata.atlassian.net/browse/CXT-718
https://cognitedata.atlassian.net/browse/CXT-699?atlOrigin=eyJpIjoiNTM1MDgyYmE1N2UwNDg5Y2IwYjQ4OTFmMmZmNjdlMWMiLCJwIjoiaiJ9

# What has changed?
## Refactoring of model structure
I heavily refactored the model structure, such that the models are now mostly represented as pure interfaces. Each interface features a discriminator field called `modelName` which is supposed to directly map to the `model.externalId` on DMS. In fact, it is actually  now also a real property on the DMS models introduced by this PR: https://github.com/cognitedata/matchmakers-property-graphs/pull/13
Besides making it possible to identify the type of objects by the modelName, this allows to selectively filter for nodes and edges of a specific model on DMS.

### Reasons for the refactoring of the model structure
In the previous version, I designed the representation of the DMS model mostly on the Adapter pattern using some generics tricks to have the upsert request create objects of the correct type. This implementation also featured a "sanitizing" step before and after upserting, since the upsert format had to differ from the retrieval format due to DMS bugs / limitations.

These limitations have been lifted in the meantime, such that the sanitizing is not necessary anymore. I therefore removed these now unnecessary steps.

In addition, I have taken a deep look into @dlcognite 's FDM SDK generator and learned a nice trick there: You can create a `ModelMap` that maps model names to types and use that as a generic in a function by giving it the model name as an argument of type `keyof ModelMap`. In this way, you can facilitate a direct coupling between a model name and its type through a generic. This is imo much more readable and cleaner to how I did it before, which was another reason for the refactoring.

## More features DMS requests
### Deletion of edges and nodes
Added the possibility to delete nodes and edges through its own function

### Handling of limits in upsert and delete requests
The upsert and delete API calls have a limit of 1 to 1000 items per request.
The new implementation properly handles this by splitting the passed items in chunks and POSTing them seperately

### Concurrent upserts and deletions
The new implementation for upsert and deletion allows for concurrent upsert and deletion (up to 10 workers) in parallel
This is done through a poor-mans task queue: The items to delete / upserts are split into chunks of up to 1000 items. 
10 worker Promises are spawned that take the first 10 chunks. When a worker with index W is finished, it spawns a worker with index W+10 as its last operation, which takes on the chunk at position W+10 next. Any finished result items are pushed into a list together with the index W of the worker. This list is sorted by W before returning to the caller, in order to preserve the order of items.

## Full upserting of the relevant parts of the GraphDocument
The relevant parts of the `GraphDocument` for the Maintain use case are now translated into nodes and edges and upserted. Those are:
- Viewbox -> Viewbox Node
- Instruments DiagramInstance -> Instrument Node 
- Valve Diagram Instance -> Valve Node
- Line Diagram Instance -> ProcessLine Node
- File Connection Diagram Instance -> FileConnection Node
- Connections between the above -> FilePageEdge Edge

This conversion is fully done in the `graphDocumentToNodesAndEdges` function in src/dms/dataModel.ts

## How to review
I suggest to review the `src/dms/dataModel.ts` first and then see how this is applied in `src/dms/request.ts`. 
The actual upserting of the nodes and edges is done in the CLI script `packages/pid-tools/cli/upsertDms.ts`. 
There is also another CLI script `packages/pid-tools/cli/listDms.ts`, which helps to list what is already there on a file, without upserting or changing anything

## How to test
Just check out the branch and run one of the shell scripts `upsert_dms.sh` or `list_dms.sh`: Those wrap the CLI scripts with already the right `tenant-id`, `client-id` and space. You need to place a valid `GraphDocument` JSON file in the given path though.